Remove warnings: 1) added #pragma warning(disable:4251) to slivr to remove dll-interface warnings (removed 386 warnings on Windows)
'size_t' to 'int' conversions (mostly when using .size() of an stl containter)

2) 'return' : conversion from 'size_t' to 'int' (cast return to 'int')
3) Change use of 'int' to 'size_t' when 'size_t' is more appropriate.
4) Change use 'size_t' to 'int' when 'int' is more appropriate
5) Cast arguments to functions from 'size_t' to 'int' when 'int' is expected.
6) use 'vtkIdType' when calling vtkDataArry->GetNumberOfTuples()
7) use 'int' when calling vtkDataArray->GetNumberOfComponents()

Reduced warnings on Windows from 6633 down to 4900.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@21610 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/DBAtts/MetaData/avtScalarMetaData.C
@@ -1038,7 +1038,7 @@ avtScalarMetaData::ComboValFromDigits(const std::vector<std::vector<int> > &ptMa
     int row, col;
     *id = 0.0;
     std::list<int>::const_reverse_iterator it;
-    for (col = digits.size()-1, it = digits.rbegin(); col >= 0; col--, it++)
+    for (col = (int)digits.size()-1, it = digits.rbegin(); col >= 0; col--, it++)
     {
         for (row = 0; row <= *it; row++)
             *id += ptMap[row][col];
