+++ src/avt/Pipeline/Data/avtSamplePoints.C
@@ -169,8 +169,8 @@ avtSamplePoints::GetNumberOfVariables(void)
 const std::string &
 avtSamplePoints::GetVariableName(int idx)
 {
-    if (idx < 0 || idx >= varnames.size())
-        EXCEPTION2(BadIndexException, idx, varnames.size());
+    if (idx < 0 || idx >= (int)varnames.size())
+        EXCEPTION2(BadIndexException, idx, (int)varnames.size());
 
     return varnames[idx];
 }
@@ -192,8 +192,8 @@ avtSamplePoints::GetVariableName(int idx)
 int
 avtSamplePoints::GetVariableSize(int idx)
 {
-    if (idx < 0 || idx >= varsize.size())
-        EXCEPTION2(BadIndexException, idx, varsize.size());
+    if (idx < 0 || idx >= (int)varsize.size())
+        EXCEPTION2(BadIndexException, idx, (int)varsize.size());
 
     return varsize[idx];
 }
