+++ src/common/comm/win32commhelpers.C
@@ -50,7 +50,7 @@ CopyHostentString(char **dest, char *src, const char *name, const char *mName)
     }
     else
     {
-        int len = strlen(src)+1;
+        size_t len = strlen(src)+1;
         *dest = (char*)malloc(len);
         memset(*dest, 0, len);
         strcpy(*dest, src);
